Put ImageViewBuildLog to separate file
